requirements.txt
# ── Core Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# ── Authentication ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ── Rate Limiting ──────────────────────────────────────────────
slowapi==0.1.9


